render-big-pucrute.js -> render-big-picture.js rename
